requirements/base.txt
Django==4.1.5
python-dotenv==0.21.0
djangorestframework==3.14.0
drf-yasg==1.21.4